Changes from review
* Keep GetPatientToPixelLUT() backwards compatible
* Change orientation methods to class methods
* Separate test fake dataset code to util module
* improve comments for dummy dose grid
* Merge decubitus tests into test_dvhcalc
